Move live runtime control into aces_runtime
