+++ implementations/python/src/aces/core/runtime/control_plane.py
@@ -1,5 +1,5 @@
 from aces._compat import reexport as _reexport
 
-_reexport(globals(), "aces_processor.control_plane")
+_reexport(globals(), "aces_runtime.control_plane")
 
 del _reexport
