+++ implementations/python/src/aces/core/runtime/control_plane_store.py
@@ -1,5 +1,5 @@
 from aces._compat import reexport as _reexport
 
-_reexport(globals(), "aces_processor.control_plane_store")
+_reexport(globals(), "aces_runtime.control_plane_store")
 
 del _reexport
