+++ implementations/python/src/aces/core/runtime/manager.py
@@ -1,5 +1,5 @@
 from aces._compat import reexport as _reexport
 
-_reexport(globals(), "aces_processor.manager")
+_reexport(globals(), "aces_runtime.manager")
 
 del _reexport
